followStatus state now updating accurately on same route re-renders
