+++ src/components/Common/FollowButton.tsx
@@ -1,4 +1,7 @@
 import { Button } from "@mui/material";
+import axios from "axios";
+import { Dispatch, SetStateAction } from "react";
+import { useAppSelector } from "../../state/hooks";
 
 const styles = {
   followButton: {
@@ -15,13 +18,37 @@ const styles = {
 };
 
 type FollowButtonProps = {
-  onClick: (e: React.MouseEvent) => void;
+  setFollowStatus: Dispatch<SetStateAction<boolean>>;
+  username: string;
 };
 
-const FollowButton = ({ onClick }: FollowButtonProps) => {
+const FollowButton = ({ setFollowStatus, username }: FollowButtonProps) => {
+  const user = useAppSelector((state) => state.user);
+
+  const handleFollow = async (e: React.MouseEvent) => {
+    e.stopPropagation();
+    try {
+      await axios.put(
+        "http://localhost:3001/api/profile/followUser",
+        {
+          userId: user.userId,
+          username,
+        },
+        {
+          headers: {
+            "Content-Type": "application/json",
+          },
+        }
+      );
+      setFollowStatus(true);
+    } catch (error) {
+      console.log(error);
+    }
+  };
+
   return (
     <Button
-      onClick={onClick}
+      onClick={handleFollow}
       size="small"
       sx={styles.followButton}
       variant="contained"
